+++ src/NumberConvertor.h
@@ -7,7 +7,7 @@ class NumberConvertor {
     NumberConvertor() = delete;
 public:
     static std::string convertHexToBinary(char hex);
-    static unsigned long long convertBinaryToLong(const std::string &binary);
+    static long convertBinaryToLong(const std::string &binary);
 };
 
 #endif
